backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==5.0.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.4
defusedxml==0.7.1
Django==4.2.4
django-celery-beat==2.8.1
django-htmx==1.23.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
Pillow==10.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2==2.9.6
psycopg2-binary==2.9.9
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytils==0.4.3
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.4.4
stripe==12.2.0
tomli==2.2.1
typing_extensions==4.7.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
